Implement reading scalar values from TEXT files.
This adds support for the TEXT DBTYPE, with LINE and COL parameters.
LINE and COL can be specified in regen4 syntax: "1-4", or using python slicing notation: "0:4".
Note with python slicing the end value exclusive, while repgen4 the end value is inclusive. Slicing is also 0-based index, vs repgen4's 1-based indexing.

Be careful when stripping quotes from input data files, and don't do that if there's an equal sign (=) present, as that could be a key-value pair. They will be appropriately stripped later on in Value.