Add a separate timezone config for DB fetches.
This splits the tz option into two, tz and dbtz.
tz is the timezone used for display of dates and times in the report output.
dbtz is the time zone used for data queries. If not set, it defaults to tz for compatibility.

Also added key=value argument support, to be more compatible with repgen4 and reducing the need to continuously add additional explicit options. Supplied arguments are passed through to the report, as if they were specified directly in the report definition.