Change TS time output to use DBTZ.
Change to use DBTZ for the report output for TS values, which makes it consistent with repgen4. Without this, data recorded in another timezone than where the report is generated will show times in the timezone of the report generation, not the timezone of the gage where the data was recorded.